requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
boto3==1.18.20
botocore==1.21.20
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.6
django-heroku==0.3.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.11.1
django-summernote==0.8.11.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
dnspython==2.1.0
gunicorn==20.1.0
graphene-django==2.15.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.6.3
incremental==21.3.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.1.0
libgravatar==1.0.0
MarkupSafe==2.0.1
Pillow==8.3.1
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycurl==7.44.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
sentry-sdk==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
twilio==6.63.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
zope.interface==5.4.0
